+++ Dapper.Tests.Performance/LegacyTests.cs
@@ -332,7 +332,7 @@ namespace Dapper.Tests.Performance
                         }
                     }, "Hand Coded");
 
-#if !COREFX
+#if !NETSTANDARD1_3
                     var table = new DataTable
                     {
                         Columns =
