+++ Dapper.Tests.Performance/Massive/Massive.cs
@@ -1,4 +1,4 @@
-﻿#if !COREFX
+﻿#if !NETSTANDARD1_3
 using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
