+++ Dapper.Tests.Performance/PetaPoco/PetaPoco.cs
@@ -1,4 +1,4 @@
-﻿#if !COREFX
+﻿#if !NETSTANDARD1_3
 using System;
 using System.Collections.Generic;
 using System.Linq;
