+++ Dapper.Tests/ConstructorTests.cs
@@ -104,28 +104,6 @@ SELECT * FROM @ExplicitConstructors"
             }
         }
 
-#if LINQ2SQL
-        private class NoDefaultConstructorWithBinary
-        {
-            public System.Data.Linq.Binary Value { get; set; }
-            public int Ynt { get; set; }
-            public NoDefaultConstructorWithBinary(System.Data.Linq.Binary val)
-            {
-                Value = val;
-            }
-        }
-
-        [Fact]
-        public void TestNoDefaultConstructorBinary()
-        {
-            byte[] orig = new byte[20];
-            new Random(123456).NextBytes(orig);
-            var input = new System.Data.Linq.Binary(orig);
-            var output = connection.Query<NoDefaultConstructorWithBinary>("select @input as val", new { input }).First().Value;
-            output.ToArray().IsSequenceEqualTo(orig);
-        }
-#endif
-
         [Fact]
         public void Issue461_TypeHandlerWorksInConstructor()
         {
