+++ Dapper.Tests/TypeHandlerTests.cs
@@ -545,7 +545,7 @@ namespace Dapper.Tests
         private static string GetDescriptionFromAttribute(MemberInfo member)
         {
             if (member == null) return null;
-#if COREFX
+#if NETCOREAPP1_0
             var data = member.CustomAttributes.FirstOrDefault(x => x.AttributeType == typeof(DescriptionAttribute));
             return (string)data?.ConstructorArguments.Single().Value;
 #else
