+++ Dapper/DynamicParameters.cs
@@ -6,7 +6,7 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Reflection.Emit;
 
-#if COREFX
+#if NETSTANDARD1_3
 using ApplicationException = System.InvalidOperationException;
 #endif
