+++ Dapper/SqlDataRecordHandler.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 
-#if !COREFX
+#if !NETSTANDARD1_3
 namespace Dapper
 {
     internal sealed class SqlDataRecordHandler : SqlMapper.ITypeHandler
