+++ Dapper/SqlMapper.TypeHandlerCache.cs
@@ -11,7 +11,7 @@ namespace Dapper
         /// </summary>
         /// <typeparam name="T">The type to have a cache for.</typeparam>
         [Obsolete(ObsoleteInternalUsageOnly, false)]
-#if !COREFX
+#if !NETSTANDARD1_3
         [Browsable(false)]
 #endif
         [EditorBrowsable(EditorBrowsableState.Never)]
