+++ Dapper/TableValuedParameter.cs
@@ -2,7 +2,7 @@
 using System.Data;
 using System.Reflection;
 
-#if !COREFX
+#if !NETSTANDARD1_3
 namespace Dapper
 {
     /// <summary>
