+++ Dapper/UdtTypeHandler.cs
@@ -5,7 +5,7 @@ namespace Dapper
 {
     public static partial class SqlMapper
     {
-#if !COREFX
+#if !NETSTANDARD1_3 && !NETSTANDARD2_0
         /// <summary>
         /// A type handler for data-types that are supported by the underlying provider, but which need
         /// a well-known UdtTypeName to be specified
